# Create a separate hook runner for onSend hooks