--- a/lib/handleRequest.js
+++ b/lib/handleRequest.js
@@ -5,7 +5,7 @@ const fastJsonStringify = require('fast-json-stringify')
 const urlUtil = require('url')
 const validation = require('./validation')
 const validateSchema = validation.validate
-const runHooks = require('./hookRunner')
+const runHooks = require('./hookRunner').hookRunner
 
 const schemas = require('./schemas.json')
 const inputSchemaError = fastJsonStringify(schemas.inputSchemaError)