--- a/lib/hookRunner.js
+++ b/lib/hookRunner.js
@@ -3,18 +3,9 @@
 function hookRunner (functions, runner, state, cb) {
   var i = 0
 
-  function next (err, value) {
-    if (err) {
+  function next (err) {
+    if (err || i === functions.length) {
       cb(err, state)
-      return
-    }
-
-    if (value !== undefined) {
-      state = value
-    }
-
-    if (i === functions.length) {
-      cb(null, state)
       return
     }
 
@@ -24,8 +15,8 @@ function hookRunner (functions, runner, state, cb) {
     }
   }
 
-  function handleResolve (value) {
-    next(null, value)
+  function handleResolve () {
+    next()
   }
 
   function handleReject (err) {
@@ -35,4 +26,39 @@ function hookRunner (functions, runner, state, cb) {
   next()
 }
 
-module.exports = hookRunner
+function onSendHookRunner (functions, reply, payload, cb) {
+  var i = 0
+
+  function next (err, newPayload) {
+    if (err) {
+      cb(err, reply, payload)
+      return
+    }
+
+    if (newPayload !== undefined) {
+      payload = newPayload
+    }
+
+    if (i === functions.length) {
+      cb(null, reply, payload)
+      return
+    }
+
+    const result = functions[i++](reply.request, reply, payload, next)
+    if (result && typeof result.then === 'function') {
+      result.then(handleResolve, handleReject)
+    }
+  }
+
+  function handleResolve (newPayload) {
+    next(null, newPayload)
+  }
+
+  function handleReject (err) {
+    cb(err, reply, payload)
+  }
+
+  next()
+}
+
+module.exports = { hookRunner, onSendHookRunner }