--- a/lib/reply.js
+++ b/lib/reply.js
@@ -8,7 +8,7 @@ const serialize = validation.serialize
 const statusCodes = require('http').STATUS_CODES
 const flatstr = require('flatstr')
 const FJS = require('fast-json-stringify')
-const runHooks = require('./hookRunner')
+const runHooks = require('./hookRunner').onSendHookRunner
 
 const serializeError = FJS({
   type: 'object',
@@ -122,24 +122,20 @@ function onSendHook (reply, payload) {
   if (reply.context.onSend !== null) {
     runHooks(
       reply.context.onSend,
-      hookIterator.bind(reply),
+      reply,
       payload,
-      wrapOnSendEnd.bind(reply)
+      wrapOnSendEnd
     )
   } else {
     onSendEnd(reply, payload)
   }
 }
 
-function hookIterator (fn, payload, next) {
-  return fn(this.request, this, payload, next)
-}
-
-function wrapOnSendEnd (err, payload) {
+function wrapOnSendEnd (err, reply, payload) {
   if (err) {
-    handleError(this, err)
+    handleError(reply, err)
   } else {
-    onSendEnd(this, payload)
+    onSendEnd(reply, payload)
   }
 }