@nwoltman 2018-02-18
It's right here: https://github.com/fastify/fastify/pull/795/files#diff-810aa058de5dd12d6a4d43556a2eb5e6R178