@mcollina 2018-02-18
I mean one using the external API.
Il giorno dom 18 feb 2018 alle 08:45 Nathan Woltman <
notifications@github.com> ha scritto:

> It's right here:
> https://github.com/fastify/fastify/pull/795/files#diff-810aa058de5dd12d6a4d43556a2eb5e6R178
>
> —
> You are receiving this because you commented.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/fastify/fastify/pull/795#issuecomment-366498426>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AADL498uoKGnfqSCz-C90q4IBgyatO2Jks5tV9UQgaJpZM4SJhi5>
> .
>